Tests refactor (#234)
* package refactor

* Update build.gradle.kts
