+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/framework/consent/payment/DefaultOBDomesticVRPConsentRequestFactory.kt
@@ -3,8 +3,8 @@ package com.forgerock.sapi.gateway.ob.uk.framework.consent.payment
 import com.forgerock.sapi.gateway.framework.configuration.requirePaymentContextCode
 import uk.org.openbanking.datamodel.v3.common.OBExternalPaymentContext1Code
 import uk.org.openbanking.datamodel.v3.vrp.OBDomesticVRPConsentRequest
-import uk.org.openbanking.testsupport.vrp.OBDomesticVrpConsentRequestTestDataFactory.aValidOBDomesticVRPConsentRequest
-import uk.org.openbanking.testsupport.vrp.OBDomesticVrpConsentRequestTestDataFactory.aValidOBDomesticVRPConsentRequestMandatoryFields
+import uk.org.openbanking.testsupport.v3.vrp.OBDomesticVrpConsentRequestTestDataFactory.aValidOBDomesticVRPConsentRequest
+import uk.org.openbanking.testsupport.v3.vrp.OBDomesticVrpConsentRequestTestDataFactory.aValidOBDomesticVRPConsentRequestMandatoryFields
 
 /**
  * Default factory implementation which is used by the functional tests OOTB.
