+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/framework/consent/payment/DefaultOBWriteDomesticConsent4Factory.kt
@@ -3,7 +3,7 @@ package com.forgerock.sapi.gateway.ob.uk.framework.consent.payment
 import com.forgerock.sapi.gateway.framework.configuration.requirePaymentContextCode
 import uk.org.openbanking.datamodel.v3.common.OBExternalPaymentContext1Code
 import uk.org.openbanking.datamodel.v3.payment.OBWriteDomesticConsent4
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory
+import uk.org.openbanking.testsupport.v3.payment.OBWriteDomesticConsentTestDataFactory
 
 /**
  * Default factory implementation which is used by the functional tests OOTB.
