+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/framework/consent/payment/DefaultOBWriteInternationalScheduledConsent5Factory.kt
@@ -3,8 +3,8 @@ package com.forgerock.sapi.gateway.ob.uk.framework.consent.payment
 import com.forgerock.sapi.gateway.framework.configuration.requirePaymentContextCode
 import uk.org.openbanking.datamodel.v3.common.OBExternalPaymentContext1Code
 import uk.org.openbanking.datamodel.v3.payment.OBWriteInternationalScheduledConsent5
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5MandatoryFields
+import uk.org.openbanking.testsupport.v3.payment.OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5
+import uk.org.openbanking.testsupport.v3.payment.OBWriteInternationalScheduledConsentTestDataFactory.aValidOBWriteInternationalScheduledConsent5MandatoryFields
 
 /**
  * Default factory implementation which is used by the functional tests OOTB.
