+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/framework/consent/payment/DefaultOBWriteInternationalStandingOrderConsent6Factory.kt
@@ -3,8 +3,8 @@ package com.forgerock.sapi.gateway.ob.uk.framework.consent.payment
 import com.forgerock.sapi.gateway.framework.configuration.requirePaymentContextCode
 import uk.org.openbanking.datamodel.v3.common.OBExternalPaymentContext1Code
 import uk.org.openbanking.datamodel.v3.payment.OBWriteInternationalStandingOrderConsent6
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalStandingOrderConsentTestDataFactory.aValidOBWriteInternationalStandingOrderConsent6
-import uk.org.openbanking.testsupport.payment.OBWriteInternationalStandingOrderConsentTestDataFactory.aValidOBWriteInternationalStandingOrderConsent6MandatoryFields
+import uk.org.openbanking.testsupport.v3.payment.OBWriteInternationalStandingOrderConsentTestDataFactory.aValidOBWriteInternationalStandingOrderConsent6
+import uk.org.openbanking.testsupport.v3.payment.OBWriteInternationalStandingOrderConsentTestDataFactory.aValidOBWriteInternationalStandingOrderConsent6MandatoryFields
 
 /**
  * Default factory implementation which is used by the functional tests OOTB.
