+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/support/payment/PaymentFactory.kt
@@ -7,7 +7,7 @@ import org.assertj.core.api.Assertions.assertThat
 import uk.org.openbanking.datamodel.v3.common.OBSupplementaryData1
 import uk.org.openbanking.datamodel.v3.payment.*
 import uk.org.openbanking.datamodel.v3.vrp.OBDomesticVRPInitiation
-import uk.org.openbanking.testsupport.payment.OBWriteFileConsentTestDataFactory
+import uk.org.openbanking.testsupport.v3.payment.OBWriteFileConsentTestDataFactory
 import java.io.File
 import java.io.StringReader
 import java.nio.charset.StandardCharsets
