+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/tests/functional/payment/domestic/payments/api/v3_1_8/GetDomesticPaymentDetails.kt
@@ -10,7 +10,7 @@ import com.forgerock.sapi.gateway.ob.uk.support.payment.PaymentFactory
 import com.forgerock.sapi.gateway.ob.uk.support.payment.defaultPaymentScopesForAccessToken
 import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion
 import uk.org.openbanking.datamodel.v3.payment.OBWritePaymentDetailsResponse1
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory
+import uk.org.openbanking.testsupport.v3.payment.OBWriteDomesticConsentTestDataFactory
 
 class GetDomesticPaymentDetails(
     val version: OBVersion,
