+++ src/main/kotlin/com/forgerock/sapi/gateway/ob/uk/tests/functional/payment/domestic/payments/consents/api/v3_1_8/GetDomesticPaymentsConsentFundsConfirmation.kt
@@ -12,7 +12,7 @@ import com.forgerock.sapi.gateway.ob.uk.support.payment.defaultPaymentScopesForA
 import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion
 import com.github.kittinunf.fuel.core.FuelError
 import uk.org.openbanking.datamodel.v3.payment.*
-import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory.aValidOBWriteDomesticConsent4
+import uk.org.openbanking.testsupport.v3.payment.OBWriteDomesticConsentTestDataFactory.aValidOBWriteDomesticConsent4
 
 class GetDomesticPaymentsConsentFundsConfirmation(
     val version: OBVersion,
